@@ -26,7 +26,6 @@ export class AccessoryComponent implements OnInit {
 
     newAccessory(): void {
         this.editAccessory = new AccessoryDecoder();
-        this.editAccessory.macro = new Macro();
         this.editAccessory.accessoryDecoderType = new DecoderAccessoryType();
     }
 
@@ -45,10 +44,7 @@ export class AccessoryComponent implements OnInit {
         newAccessory.address = accessory.address;
         newAccessory.name = accessory.name;
         newAccessory.currentValue = accessory.currentValue;
-        newAccessory.macro = accessory.macro;
-        if (newAccessory.macro == null) {
-            newAccessory.macro = new Macro();
-        }
+        newAccessory.macroId = accessory.macroId;
         this.setCurrentAccessory(newAccessory);
     }
 
@@ -57,9 +53,6 @@ export class AccessoryComponent implements OnInit {
     }
 
     saveAccessory(): void {
-        if (this.editAccessory.macro.macroId == null || this.editAccessory.macro.macroId.toString() == "") {
-            this.editAccessory.macro = null;
-        }
         this.accessoryDecoderService.saveAccessory(this.editAccessory);
         this.newAccessory();
     }
