@@ -1,10 +1,9 @@
 /**
  * Created by shawn on 03/04/17.
  */
-import {Macro} from "./macro";
 export class Block {
     blockId: string;
     blockName: string;
     occupied: boolean;
-    macro: Macro;
+    macroId: number;
 }
