@@ -2,9 +2,10 @@ package uk.co.redkiteweb.dccweb;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -21,6 +22,7 @@ import javax.sql.DataSource;
 public class DccWebApplication {
 
     @Bean
+    @Primary
     @ConfigurationProperties(prefix = "datasource.dccweb")
     public DataSource dataSource() {
         return DataSourceBuilder.create().build();
