@@ -4,8 +4,8 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-import uk.co.redkiteweb.dccweb.data.data.LogEntry;
-import uk.co.redkiteweb.dccweb.data.store.LogStore;
+import uk.co.redkiteweb.dccweb.data.LogEntry;
+import uk.co.redkiteweb.dccweb.store.LogStore;
 
 import java.util.ArrayList;
 
@@ -31,7 +31,7 @@ public class LogEntryMessagesTest {
 
     @Test
     public void testGetMessages() {
-        when(logStore.getLastSix()).thenReturn(new ArrayList<LogEntry>());
+        when(logStore.getLastSix()).thenReturn(new ArrayList<>());
         assertNotNull(logMessages.getMessages());
     }
 }
