@@ -3,8 +3,7 @@ package uk.co.redkiteweb.dccweb.service;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.embedded.EmbeddedServletContainerInitializedEvent;
-import org.springframework.context.ApplicationListener;
+import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import uk.co.redkiteweb.dccweb.model.RegisterDetails;
@@ -15,13 +14,19 @@ import java.net.Inet4Address;
 import java.net.UnknownHostException;
 
 @Service
-public class RelayService implements ApplicationListener<EmbeddedServletContainerInitializedEvent> {
+public class RelayService {
 
     private static final Logger LOGGER = LogManager.getLogger(RelayService.class);
 
     private RestTemplate restTemplate;
-    private Integer port;
     private RelayControllerStore relayControllerStore;
+    private Environment environment;
+
+    @Autowired
+    public void setEnvironment(final Environment environment) {
+        this.environment = environment;
+
+    }
 
     @Autowired
     public void setRelayControllerStore(final RelayControllerStore relayControllerStore) {
@@ -33,11 +38,6 @@ public class RelayService implements ApplicationListener<EmbeddedServletContaine
         this.restTemplate = restTemplate;
     }
 
-    @Override
-    public void onApplicationEvent(final EmbeddedServletContainerInitializedEvent embeddedServletContainerInitializedEvent) {
-        port = embeddedServletContainerInitializedEvent.getEmbeddedServletContainer().getPort();
-    }
-
     public void registerWithServer(final RegisterDetails registerDetails) {
         final String url = String.format("http://%s:%d/api/relay-controller/update",
                 registerDetails.getHost(), registerDetails.getPort());
@@ -45,7 +45,7 @@ public class RelayService implements ApplicationListener<EmbeddedServletContaine
             final String host = Inet4Address.getLocalHost().getHostAddress();
             final RelayController relayController = relayControllerStore.getRelayController();
             relayController.setControllerId("DUMMY");
-            relayController.setIpAddress(String.format("%s:%d", host, port));
+            relayController.setIpAddress(String.format("%s:%s", host, environment.getProperty("local.server.port")));
             relayControllerStore.setRelayController(relayController);
             relayControllerStore.setValue(restTemplate.postForEntity(url, relayController, Integer.class).getBody());
         } catch (final UnknownHostException exception) {
